+++ src/meico/app/gui/WebBrowser.java
@@ -32,7 +32,7 @@ public class WebBrowser extends StackPane {
 
             this.webEngine = browser.getEngine();
             this.webEngine.setJavaScriptEnabled(true);                                      // JavaScript should be enabled by default, this is just to be sure that it is
-            this.getChildren().addAll(browser);
+            this.getChildren().add(browser);
             this.openURL("https://www.verovio.org/index.xhtml");                            // set a start page
         }
     }
@@ -85,5 +85,4 @@ public class WebBrowser extends StackPane {
             return false;
         }
     }
-
 }